requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.5
aiosignal==1.3.1
aiostream==0.5.2
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
exceptiongroup==1.1.3
frozenlist==1.4.0
fsspec==2023.10.0
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.327
langsmith==0.0.54
llama-index==0.8.56
magic-filter==1.0.11
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numpy==1.26.1
openai==0.28.1
packaging==23.2
pandas==2.1.2
pydantic==1.10.13
pypdf==3.17.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.5
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wrapt==1.15.0
yarl==1.9.2
